(WIP) Moved database package under commons.
