+++ app/src/androidTest/java/org/hisp/dhis/android/dataentry/DhisInstrumentationTestsApp.java
@@ -5,7 +5,7 @@ import android.support.test.espresso.Espresso;
 import android.support.test.espresso.IdlingResource;
 
 import org.hisp.dhis.android.core.configuration.ConfigurationModel;
-import org.hisp.dhis.android.dataentry.database.DbModule;
+import org.hisp.dhis.android.dataentry.commons.database.DbModule;
 import org.hisp.dhis.android.dataentry.server.ServerComponent;
 import org.hisp.dhis.android.dataentry.utils.IdlingSchedulerProvider;
 import org.hisp.dhis.android.dataentry.utils.SchedulerModule;
